+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.3.1]
+- Added REPL start message
+- Added loading Neko files in REPL (.load)
+- Added saving session history of evaluated lines in REPL (.save)
+- Added Editor mode to repl (.editor)
+
 ## [0.3.0]
 - Made improvements to CLI and REPL
 - Added comparison operations